# perf: improve performance of path_string impl for EvaluationContext